+++ Justsurviveoffice/src/main/java/com/spring/app/users/domain/CategoryDTO.java
@@ -17,7 +17,7 @@ import lombok.Setter;
 @Builder              // 생성자 대신, 필요한 값만 선택해서 체이닝 방식으로 객체를 만들 수 있게 해주는 것.
 public class CategoryDTO {
 	
-	private int categoryNo;
+	private Long categoryNo;
 	
 	private String categoryName;
